Address remaining code review feedback
- Fix escaped footnote markers in rust-testing-with-rstest-fixtures.md
  (change \[^n\] to [^n] format)
- Remove unused footnote definitions [^6], [^14], [^23]
- Extract frontmatter tests from src/process.rs to new
  tests/process_frontmatter.rs to bring process.rs under 400 lines

Note: frontmatter module visibility kept as pub with #[doc(hidden)]
rather than pub(crate) because the binary crate requires access to
split_leading_yaml_frontmatter.
